.ipynb_checkpoints/requirements-checkpoint.txt
mcp
numpy
Pillow
google-generativeai
mcp
ultralytics

